sitemap: remove filenamefetcher
